requirements.install.txt
pyyaml~=5.2
send2trash~=1.5.0
google-oauth2l~=0.9.0
oauth2client~=2.1.0
requests~=2.19.1
